Restructures FacadeTests/. Removes Tests from the folder name.
