+++ Tests/Unit/ApiVersionTest.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace LoRaWan.Tests.Unit.CommonAPI
+namespace LoRaWan.Tests.Unit.LoRaTools.CommonAPI
 {
     using System;
     using System.Collections.Generic;
